+++ changelog.md
@@ -9,6 +9,7 @@ v2.0.6
 * Update textures for cloth armor
 * Allow re-mapping of grimoire navigation keybinds
 * Make right-clicking in the grimoire GUI go back a topic
+* Add four-leaf clovers, to help you concoct some luck if you can find one in the tall grass
 
 v2.0.5
 ------